pyproject.toml
[tool.poetry]
name = "kite-services"
version = "1.0.0"
description = "Kite Trading Services - Consolidated API for stock market data and analysis"
authors = ["Ashok Kumar <ashok@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "0.104.1"
uvicorn = {extras = ["standard"], version = "0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
kiteconnect = "^4.2.0"
yfinance = "^0.2.28"
yahooquery = "^2.3.7"
sqlalchemy = "^2.0.23"
alembic = "^1.13.1"
aiosqlite = "^0.19.0"
httpx = "^0.25.2"
aiohttp = "^3.9.1"
pandas = "^2.2.0"
numpy = "^1.26.0"
websockets = "^14.0"
structlog = "^23.2.0"
redis = "^5.0.1"
python-dotenv = "^1.0.0"
pyyaml = "^6.0.1"
python-dateutil = "^2.8.2"
pytz = "^2023.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"

[tool.poetry.scripts]
kite-services = "main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = "-v --tb=short"
markers = [
    "smoke: Smoke tests that must pass",
    "integration: Integration tests",
    "e2e: End-to-end tests",
    "production: Production-specific tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
